pyproject.toml
[project]
name = "snowpark-widf-lambda-loader"
version = "0.1.0"
description = "AWS Lambda function that loads data from S3 to Snowflake using Workload Identity Federation (WIDF)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "snowflake-snowpark-python[pandas]>=1.44.0",
    "boto3>=1.34.0",
]

[project.scripts]
widf-loader = "app:main"
